+++ src/main/java/com/gk_dreams/HLuxe/service/RoomServiceImpl.java
@@ -3,13 +3,16 @@ package com.gk_dreams.HLuxe.service;
 import com.gk_dreams.HLuxe.dto.RoomDto;
 import com.gk_dreams.HLuxe.entity.Hotel;
 import com.gk_dreams.HLuxe.entity.Room;
+import com.gk_dreams.HLuxe.entity.User;
 import com.gk_dreams.HLuxe.exceptions.ResourceNotFoundException;
+import com.gk_dreams.HLuxe.exceptions.UnAuthorisedException;
 import com.gk_dreams.HLuxe.repository.HotelRepository;
 import com.gk_dreams.HLuxe.repository.InventoryRepository;
 import com.gk_dreams.HLuxe.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,6 +34,11 @@ public class RoomServiceImpl implements RoomService{
     public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(
                 ()-> new ResourceNotFoundException("Hotel Not found with Id: "+hotelId));
+
+        if(!getCurrentUser().equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own hotel with Id: "+ hotelId);
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -41,7 +49,7 @@ public class RoomServiceImpl implements RoomService{
     }
 
     @Override
-    public List<RoomDto> getALlRooms(Long hotelId) {
+    public List<RoomDto> getAllRooms(Long hotelId) {
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(
                 ()-> new ResourceNotFoundException("Hotel Not found with Id: "+hotelId));
         return hotel.getRooms()
@@ -62,7 +70,16 @@ public class RoomServiceImpl implements RoomService{
     public void deleteRoom(Long roomId) {
         Room room = roomRepository.findById(roomId).orElseThrow(
                 ()-> new ResourceNotFoundException("Room Not found with Id: "+roomId));
+
+        if(!getCurrentUser().equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user does not own room with Id: "+ roomId);
+        }
+
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
+    }
+
+    private User getCurrentUser(){
+        return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 }
